+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_multiapi.py
@@ -202,7 +202,7 @@ class TestMultiapi(FormRecognizerTest):
         # test that the addition of new attributes in v2023-07-31 does not break v2022-08-31
 
         with pytest.raises(ValueError) as excinfo:
-            client.begin_analyze_document("prebuilt-layout", my_file, features=[AnalysisFeature.OCR_FONT])
+            client.begin_analyze_document("prebuilt-layout", my_file, features=[AnalysisFeature.STYLE_FONT])
         assert "Keyword argument 'features' is only available for API version V2023_07_31 and later." == str(excinfo.value)
 
         # test that the addition of new methods in v2023-07-31 does not break v2022-08-31