+++ sdk/formrecognizer/azure-ai-formrecognizer/tests/test_multiapi_async.py
@@ -206,7 +206,7 @@ class TestMultiapi(AsyncFormRecognizerTest):
             # test that the addition of new attributes in v2023-07-31 does not break v2022-08-31
 
             with pytest.raises(ValueError) as excinfo:
-                await client.begin_analyze_document("prebuilt-layout", my_file, features=[AnalysisFeature.OCR_FONT])
+                await client.begin_analyze_document("prebuilt-layout", my_file, features=[AnalysisFeature.STYLE_FONT])
             assert "Keyword argument 'features' is only available for API version V2023_07_31 and later." == str(excinfo.value)
 
             # test that the addition of new methods in v2023-07-31 does not break v2022-08-31