+++ sdk/formrecognizer/azure-ai-formrecognizer/CHANGELOG.md
@@ -13,6 +13,7 @@ This version of the client library defaults to the service API version `2023-07-
 - Removed `images` property from `DocumentPage`.
 - Removed `DocumentImage` model.
 - Removed `common_name` property from `DocumentKeyValuePair`.
+- Changed `AnalysisFeature` enum member names and values. Supported enum members are: `OCR_HIGH_RESOLUTION`, `LANGUAGES`, `BARCODES`, `FORMULAS`, `KEY_VALUE_PAIRS`, `STYLE_FONT`.
 
 ### Bugs Fixed